+++ drivers/CMakeLists.txt
@@ -1,3 +1,4 @@
 add_subdirectory(faux)
+add_subdirectory(sbet)
 
 set(PDAL_TARGET_OBJECTS ${PDAL_TARGET_OBJECTS} PARENT_SCOPE)
