+++ drivers/faux/CMakeLists.txt
@@ -16,4 +16,5 @@ set(incs
     FauxReader.hpp
 )
 
-PDAL_ADD_DRIVER(reader faux "${srcs}" "${incs}")
+PDAL_ADD_DRIVER(reader faux "${srcs}" "${incs}" objects)
+set(PDAL_TARGET_OBJECTS ${PDAL_TARGET_OBJECTS} ${objects} PARENT_SCOPE)
