+++ drivers/sbet/SbetCommon.hpp
@@ -34,47 +34,11 @@
 
 #pragma once
 
-#include <pdal/IStream.hpp>
-#include <pdal/PointBuffer.hpp>
-#include <pdal/Reader.hpp>
-#include <pdal/drivers/sbet/Common.hpp>
+#include <pdal/Dimension.hpp>
 
 namespace pdal
 {
-namespace drivers
-{
-namespace sbet
-{
 
-class PDAL_DLL SbetReader : public pdal::Reader
-{
-public:
-    SET_STAGE_NAME("drivers.sbet.reader", "SBET Reader")
-    SET_STAGE_LINK("http://pdal.io/stages/drivers.sbet.reader.html")
-    SET_STAGE_ENABLED(true)
+PDAL_DLL Dimension::IdList fileDimensions();
 
-    SbetReader() : Reader()
-        {}
-
-    static Options getDefaultOptions();
-    static Dimension::IdList getDefaultDimensions()
-        { return fileDimensions(); }
-
-private:
-    std::unique_ptr<ILeStream> m_stream;
-    // Number of points in the file.
-    point_count_t m_numPts;
-    point_count_t m_index;
-
-    virtual void addDimensions(PointContextRef ctx);
-    virtual void ready(PointContextRef ctx);
-    virtual point_count_t read(PointBuffer& buf, point_count_t count);
-    virtual bool eof();
-
-    void seek(PointId idx);
-};
-
-} // namespace sbet
-} // namespace drivers
 } // namespace pdal
-
