+++ drivers/sbet/SbetWriter.cpp
@@ -32,77 +32,45 @@
 * OF SUCH DAMAGE.
 ****************************************************************************/
 
-#include <pdal/drivers/sbet/Reader.hpp>
+#include <SbetWriter.hpp>
+
+#include <SbetCommon.hpp>
+#include <pdal/PointBuffer.hpp>
 
 namespace pdal
 {
-namespace drivers
-{
-namespace sbet
-{
 
-Options SbetReader::getDefaultOptions()
+void SbetWriter::processOptions(const Options& options)
 {
-    Options options;
-    return options;
+    m_filename = options.getOption("filename").getValue<std::string>();
 }
 
 
-void SbetReader::addDimensions(PointContextRef ctx)
+void SbetWriter::ready(PointContextRef ctx)
 {
-    ctx.registerDims(getDefaultDimensions());
+    m_stream.reset(new OLeStream(m_filename));
 }
 
 
-void SbetReader::ready(PointContextRef ctx)
+void SbetWriter::write(const PointBuffer& buf)
 {
-    size_t fileSize = FileUtils::fileSize(m_filename);
-    size_t pointSize = getDefaultDimensions().size() * sizeof(double);
-    if (fileSize % pointSize != 0)
-        throw pdal_error("invalid sbet file size");
-    m_numPts = fileSize / pointSize;
-    m_index = 0;
-    m_stream.reset(new ILeStream(m_filename));
-}
+    m_callback->setTotal(buf.size());
+    m_callback->invoke(0);
 
-
-point_count_t SbetReader::read(PointBuffer& buf, point_count_t count)
-{
-    PointId nextId = buf.size();
-    PointId idx = m_index;
-    point_count_t numRead = 0;
-    seek(idx);
     Dimension::IdList dims = getDefaultDimensions();
-    while (numRead < count && idx < m_numPts)
+    for (PointId idx = 0; idx < buf.size(); ++idx)
     {
         for (auto di = dims.begin(); di != dims.end(); ++di)
         {
-            double d;
-            *m_stream >> d;
+            // If a dimension doesn't exist, write 0.
             Dimension::Id::Enum dim = *di;
-            buf.setField(dim, nextId, d);
+            *m_stream << (buf.hasDim(dim) ?
+                buf.getFieldAs<double>(dim, idx) : 0.0);
         }
-        idx++;
-        nextId++;
-        numRead++;
+        if (idx % 100 == 0)
+            m_callback->invoke(idx + 1);
     }
-    m_index = idx;
-    return numRead;
+    m_callback->invoke(buf.size());
 }
 
-
-bool SbetReader::eof()
-{
-    return m_index >= m_numPts;
-}
-
-
-void SbetReader::seek(PointId idx)
-{
-    m_stream->seek(idx * sizeof(double) * getDefaultDimensions().size());
-}
-
-} // namespace sbet
-} // namespace drivers
 } // namespace pdal
-
