+++ test/data/sbet/pipeline.xml
@@ -2,7 +2,7 @@
 <Pipeline version="1.0">
   <Writer type="drivers.text.writer">
     <Option name="filename">outfile.txt</Option>
-    <Reader type="drivers.sbet.reader">
+    <Reader type="readers.sbet">
       <Option name="filename">2-points.sbet</Option>
     </Reader>
   </Writer>
