+++ test/unit/CMakeLists.txt
@@ -121,6 +121,7 @@ include_directories(
     ${GDAL_INCLUDE_DIR}
     ${GEOTIFF_INCLUDE_DIR}
     ${PROJECT_SOURCE_DIR}/drivers/faux
+    ${PROJECT_SOURCE_DIR}/drivers/sbet
 )
 set(deps ${PDAL_LIB_NAME})
 PDAL_ADD_TEST(pdal_test "${PDAL_UNITTEST_SRCS}" "${deps}")
